refactor: throw exception in `TaskService.deleteTask()` when task not found.

+++ Oskar-Swiatek/ProjectManagerApp/src/main/java/org/example/projectmanagerapp/service/TaskService.java
@@ -38,6 +38,9 @@ public class TaskService {
     }
 
     public void deleteTask(Long id) {
+        if (taskRepository.findById(id).isEmpty()) {
+            throw new RuntimeException("Task not found");
+        }
         taskRepository.deleteById(id);
     }
 }
